Added a utility for reading timestamps produced by this utility.
NOTE: We haven't found a way to read the timestamp produced by the software Point Grey provides just yet because it appears they store the timestamp before converting the image.

For images to be produced that this module can read, you need to use a .dll produced by the latest C code (which will be pushed shortly).